% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cartera_proyectos_data}
\alias{cartera_proyectos_data}
\title{Sustainable Investment Project Portfolio (Local Data)}
\format{
A data frame (tibble).
}
\source{
\url{https://datosabiertos.senace.gob.pe}
}
\usage{
cartera_proyectos_data
}
\description{
A bundled dataset containing SENACE sustainable investment projects and their
status.
}
\keyword{datasets}
