% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{consultoras_ambientales_data}
\alias{consultoras_ambientales_data}
\title{Environmental Consulting Firms Register (Local Data)}
\format{
A data frame (tibble).
}
\source{
\url{https://datosabiertos.senace.gob.pe}
}
\usage{
consultoras_ambientales_data
}
\description{
A bundled dataset of environmental consulting firms authorized by SENACE.
}
\keyword{datasets}
