% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gasto_fuente_data}
\alias{gasto_fuente_data}
\title{Expenditure by Funding Source (Local Data)}
\format{
A data frame (tibble).
}
\source{
\url{https://datosabiertos.senace.gob.pe}
}
\usage{
gasto_fuente_data
}
\description{
A bundled dataset of budget expenditure by funding source.
}
\keyword{datasets}
