% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gasto_generica_data}
\alias{gasto_generica_data}
\title{Expenditure by Generic Classifier (Local Data)}
\format{
A data frame (tibble).
}
\source{
\url{https://datosabiertos.senace.gob.pe}
}
\usage{
gasto_generica_data
}
\description{
A bundled dataset of budget expenditure by generic classifier.
}
\keyword{datasets}
