% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reclamos_data}
\alias{reclamos_data}
\title{Complaints and Claims Register (Local Data)}
\format{
A data frame (tibble).
}
\source{
\url{https://datosabiertos.senace.gob.pe}
}
\usage{
reclamos_data
}
\description{
A bundled dataset of complaints and claims submitted to SENACE.
}
\keyword{datasets}
