% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{senace_api_key}
\alias{senace_api_key}
\title{Retrieve the active SENACE API key}
\usage{
senace_api_key()
}
\value{
A character string containing the SENACE API key.
}
\description{
Looks for an API key in the \code{SENACE_API_KEY} environment variable and then
in the \code{senace.api_key} global option. An informative error is raised if no
key is available.
}
