% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{senace_gastos_fuente}
\alias{senace_gastos_fuente}
\title{Retrieve expenditure by funding source}
\usage{
senace_gastos_fuente(api_key = NULL)
}
\arguments{
\item{api_key}{A SENACE API key. If \code{NULL}, it is retrieved automatically.}
}
\value{
A tibble with budget expenditure records summarized by funding source.
}
\description{
Downloads budget expenditure summarized by funding source from the REST API.
}
\examples{
\dontrun{
senace_gastos_fuente()
}
}
