% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{senace_gastos_generica}
\alias{senace_gastos_generica}
\title{Retrieve expenditure by generic classifier}
\usage{
senace_gastos_generica(api_key = NULL)
}
\arguments{
\item{api_key}{A SENACE API key. If \code{NULL}, it is retrieved automatically.}
}
\value{
A tibble with budget expenditure records grouped by generic classifier.
}
\description{
Downloads budget expenditure by generic classifier from the REST API.
}
\examples{
\dontrun{
senace_gastos_generica()
}
}
