% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{senace_set_key}
\alias{senace_set_key}
\title{Set the SENACE API key}
\usage{
senace_set_key(key)
}
\arguments{
\item{key}{A non-empty character string containing a personal API key.}
}
\value{
Invisibly, \code{NULL}.
}
\description{
Stores a SENACE API key in the environment of the current R session.
}
\examples{
senace_set_key("my_api_key")
}
