% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solicitud_acceso_data}
\alias{solicitud_acceso_data}
\title{Public Information Access Requests (Local Data)}
\format{
A data frame (tibble).
}
\source{
\url{https://datosabiertos.senace.gob.pe}
}
\usage{
solicitud_acceso_data
}
\description{
A bundled dataset of public information access requests.
}
\keyword{datasets}
