% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{visitas_data}
\alias{visitas_data}
\title{Institutional Visits Register (Local Data)}
\format{
A data frame (tibble).
}
\source{
\url{https://datosabiertos.senace.gob.pe}
}
\usage{
visitas_data
}
\description{
A bundled dataset of visits to SENACE offices.
}
\keyword{datasets}
