% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{senace_cartera_proyectos}
\alias{senace_cartera_proyectos}
\title{Retrieve the Sustainable Investment Project Portfolio}
\usage{
senace_cartera_proyectos(api_key = NULL)
}
\arguments{
\item{api_key}{A SENACE API key. If \code{NULL}, it is retrieved automatically.}
}
\value{
A tibble.
}
\description{
Downloads the SENACE project portfolio from the REST API.
}
\examples{
\dontrun{
senace_cartera_proyectos()
}
}
