% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{senace_consultoras_ambientales}
\alias{senace_consultoras_ambientales}
\title{Retrieve the Environmental Consulting Firms Register}
\usage{
senace_consultoras_ambientales(api_key = NULL, offline = FALSE)
}
\arguments{
\item{api_key}{A SENACE API key. If \code{NULL}, it is retrieved automatically.}

\item{offline}{If \code{TRUE}, use the dataset bundled with the package instead of
querying the API.}
}
\value{
A tibble.
}
\description{
Downloads the register of environmental consulting firms authorized by SENACE
or retrieves its bundled local snapshot.
}
