% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{senace_gastos_especifica}
\alias{senace_gastos_especifica}
\title{Retrieve expenditure by detailed classifier}
\usage{
senace_gastos_especifica(api_key = NULL, offline = FALSE)
}
\arguments{
\item{api_key}{A SENACE API key. If \code{NULL}, it is retrieved automatically.}

\item{offline}{If \code{TRUE}, use the dataset bundled with the package instead of
querying the API.}
}
\value{
A tibble.
}
\description{
Downloads detailed budget expenditure data or retrieves its bundled local
snapshot.
}
