% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{senace_gastos_especifica}
\alias{senace_gastos_especifica}
\title{Retrieve expenditure by detailed classifier}
\usage{
senace_gastos_especifica(api_key = NULL)
}
\arguments{
\item{api_key}{A SENACE API key. If \code{NULL}, it is retrieved automatically.}
}
\value{
A tibble.
}
\description{
Downloads detailed budget expenditure data from the REST API.
}
\examples{
\dontrun{
senace_gastos_especifica()
}
}
