% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{senace_reclamos}
\alias{senace_reclamos}
\title{Retrieve the Complaints and Claims Register}
\usage{
senace_reclamos(api_key = NULL, offline = FALSE)
}
\arguments{
\item{api_key}{A SENACE API key. If \code{NULL}, it is retrieved automatically.}

\item{offline}{If \code{TRUE}, use the dataset bundled with the package instead of
querying the API.}
}
\value{
A tibble.
}
\description{
Downloads complaints and claims submitted by the public or retrieves their
bundled local snapshot.
}
\examples{
senace_reclamos(offline = TRUE)
}
