% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{senace_reclamos}
\alias{senace_reclamos}
\title{Retrieve the Complaints and Claims Register}
\usage{
senace_reclamos(api_key = NULL)
}
\arguments{
\item{api_key}{A SENACE API key. If \code{NULL}, it is retrieved automatically.}
}
\value{
A tibble.
}
\description{
Downloads complaints and claims submitted by the public from the REST API.
}
\examples{
\dontrun{
senace_reclamos()
}
}
