% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{senace_solicitudes_acceso}
\alias{senace_solicitudes_acceso}
\title{Retrieve public information access requests}
\usage{
senace_solicitudes_acceso(api_key = NULL, offline = FALSE)
}
\arguments{
\item{api_key}{A SENACE API key. If \code{NULL}, it is retrieved automatically.}

\item{offline}{If \code{TRUE}, use the dataset bundled with the package instead of
querying the API.}
}
\value{
A tibble.
}
\description{
Downloads public information access requests or retrieves their bundled local
snapshot.
}
