% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{senace_solicitudes_acceso}
\alias{senace_solicitudes_acceso}
\title{Retrieve public information access requests}
\usage{
senace_solicitudes_acceso(api_key = NULL)
}
\arguments{
\item{api_key}{A SENACE API key. If \code{NULL}, it is retrieved automatically.}
}
\value{
A tibble.
}
\description{
Downloads public information access requests from the REST API.
}
\examples{
\dontrun{
senace_solicitudes_acceso()
}
}
