% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{senace_visitas}
\alias{senace_visitas}
\title{Retrieve the Institutional Visits Register}
\usage{
senace_visitas(api_key = NULL, offline = FALSE)
}
\arguments{
\item{api_key}{A SENACE API key. If \code{NULL}, it is retrieved automatically.}

\item{offline}{If \code{TRUE}, use the dataset bundled with the package instead of
querying the API.}
}
\value{
A tibble.
}
\description{
Downloads records of visits to SENACE offices or retrieves their bundled local
snapshot.
}
\examples{
senace_visitas(offline = TRUE)
}
